rename ./internal to ./interna
