+++ .gitignore
@@ -1,0 +1,2 @@
+## Directory-based project format
+.idea/
