+++ interna/dic/dic.go
