+++ interna/dic/doc.go
