+++ interna/dic/morph.go
