+++ interna/dic/trie.go
