+++ interna/dic/udic.go
