+++ interna/lattice/doc.go
