+++ interna/lattice/node.go
