+++ interna/lattice/node_test.go
