+++ tokenizer/token_test.go
@@ -19,8 +19,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/ikawaha/kagome/internal/dic"
-	"github.com/ikawaha/kagome/internal/lattice"
+	"github.com/ikawaha/kagome/interna/dic"
+	"github.com/ikawaha/kagome/interna/lattice"
 )
 
 func TestTokenClassString(t *testing.T) {
